\name{indo}
\docType{data}
\alias{indo}
\title{Formant values for Indonesian vowels}
\description{
This data set gives F1 and F2 values for five vowels of Standard Indonesian, as spoken by eight speakers (4 male, 4 female), measured from wordlist recordings.
}
\usage{indo}
\format{A data frame of 1725 rows, with columns \dQuote{subj}, \dQuote{gender}, \dQuote{vowel}, \dQuote{f1}, and \dQuote{f2}.}
\keyword{datasets}
