\name{pretty.ticks}
\title{Calculate nice tick locations given the extrema of the data.}
\description{Calculates tick positions that are \dQuote{nice round numbers} in
			 that they are always factors of 10, 5, or 2.}
\usage{
    pretty.ticks(lim)
}
\arguments{
    \item{lim}{2-element numeric vector of data extrema.}
}
\value{A numeric vector of tick locations.}
\author{McCloy, Daniel \email{drmccloy@uw.edu}}
\references{}
\seealso{\code{\link{plot.vowels}}}
\examples{}
\keyword{}
