\name{indoVowels}
\docType{data}
\alias{indo}
\title{Formant frequency measurements for Indonesian vowels}
\description{This data set gives F1 and F2 values for five vowels of Standard 
             Indonesian, as spoken by eight speakers (4 male, 4 female), 
             measured from wordlist recordings.}
\usage{indo}
\format{A data frame of 1725 rows, with columns \code{subj}, \code{gender}, \code{vowel}, \code{f1}, and \code{f2}.}
\source{
McCloy, D. R. 2014 \dQuote{Phonetic effects of morphological structure in Indonesian vowel reduction}. \emph{Proceedings of Meetings on Acoustics} 12, 060009. \url{http://dx.doi.org/10.1121/1.4870068}
}
\keyword{datasets}
